# Conflicts:
#	Motor2D/Link.cpp

+++ Motor2D/Entity.cpp
@@ -42,7 +42,7 @@ bool Entity::GotHit(Entity *& entity, Ability *& ability, Spell* &spell)
 }
 
 
-Ability* Entity::AddAbility(int number, int cooldow, int base_damage, int damage_multiplier, char* name)
+Ability* Entity::AddAbility(int number, int cooldow, int base_damage, float damage_multiplier, char* name)
 {
 	Ability* ret = nullptr;
