+++ Motor2D/MageSkeleton.cpp
@@ -84,12 +84,12 @@ bool MageSkeleton::Update(float dt)
 		{
 			if (spell != nullptr)
 			{
-				DealDamage((entity->stats.power * spell->stats.damage_multiplicator) + ability->damage); // Spells control their own damage mutiplicator
+				DealDamage(((float)entity->stats.power * (float)spell->stats.damage_multiplicator) + (float)ability->damage); // Spells control their own damage mutiplicator
 
 				spell->Effects(entity, this, ability);
 			}
 			else
-				DealDamage((entity->stats.power * ability->damage_multiplicator) + ability->damage);
+				DealDamage(((float)entity->stats.power * (float)ability->damage_multiplicator) + (float)ability->damage);
 
 			if (state == MSkl_S_Idle)
 			{
